refactor: simplify app structure and consolidate auth functionality
- Remove MainLayout component and flatten page structure
- Move Navbar and ProfileDropdown to /ui directory for better organization
- Consolidate all auth-related code into /auth directory:
  * Move axios-instance.ts → auth/api.ts
  * Move ProtectedRoute.tsx → auth/ProtectedRoute.tsx
  * Move useAuth0Integration.ts → auth/useAuth0Integration.ts
  * Move auth-types.ts → auth/types.ts
- Delete EditProfilePage and remove Edit Profile menu item
- Implement authentication-based routing:
  * Root path redirects to /home for authenticated users
  * Unauthenticated users redirected to /login
  * ProtectedRoute redirects to /login instead of inline form
- Optimize authentication loading:
  * Make token retrieval non-blocking
  * Improve loading state management
  * Reduce "Authenticating" wait times
- Add Navbar to all pages for consistent navigation
- Update all import paths throughout codebase
- Clean up empty directories and remove unused files

This refactoring significantly simplifies the app structure while improving
user experience and code maintainability.
