+++ client/src/App.tsx
@@ -2,8 +2,8 @@ import useAppInit from "./hooks/useAppInit";
 import Loading from "./components/shared/Loading";
 import { RouterProvider } from "react-router-dom";
 import { router } from "./router";
-import { useAuth0Integration } from "@/hooks/useAuth0Integration";
-import { setTokenGetter } from "./api/axios-instance";
+import { useAuth0Integration } from "@/auth/useAuth0Integration";
+import { setTokenGetter } from "./auth/api";
 import { useEffect } from "react";
 
 function App() {
