--- client/src/components/layout/Footer.tsx
@@ -1,7 +1,0 @@
-const Footer = () => {
-  return (
-    <div>Footer</div>
-  )
-}
-
-export default Footer
