--- client/src/components/layout/Header.tsx
@@ -1,7 +1,0 @@
-const Header = () => {
-  return (
-    <div>Header</div>
-  )
-}
-
-export default Header
