+++ client/src/features/slideshow-form/components/experience/AddExperienceButton.tsx
@@ -1,6 +1,6 @@
 import { Box, HStack } from "@chakra-ui/react";
 import { FiPlus } from "react-icons/fi";
-import BaseButton from "@/components/ui/BaseButton";
+import BaseButton from "@/components/shared/BaseButton";
 
 interface AddExperienceButtonProps {
   onAdd: () => void;
