--- client/src/pages/About.tsx
@@ -1,15 +1,0 @@
-import { Container, Text } from "@chakra-ui/react";
-
-const About = () => {
-  return (
-    <>
-    <div>
-      <Container>
-        <Text>About</Text>
-      </Container>
-    </div>
-    </>
-  );
-};
-
-export default About;
