--- client/src/pages/index.ts
@@ -1,3 +1,0 @@
-export { default as Home } from './Home';
-export { default as About } from './About';
-export { default as EditProfilePage } from './EditProfilePage';
