--- client/src/styles/global.css
@@ -1,3 +1,0 @@
-body {
-  background-color: blanchedalmond;
-}
