+++ client/src/ui/index.ts
@@ -1,0 +1,5 @@
+export { default as Home } from './Home';
+export { default as About } from './About';
+export { default as Login } from './Login';
+export { default as Navbar } from './Navbar';
+export { default as ProfileDropdown } from './ProfileDropdown';
